+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Astrometry/TargetCircumstances.cs
@@ -29,7 +29,7 @@ namespace Assistant.NINAPlugin.Astrometry {
             this.endTime = twilightSpan.EndTime;
 
             string cacheKey = GetCacheKey();
-            TSLogger.Trace($"TargetCircumstances cache key: {cacheKey}");
+            TSLogger.Debug($"TargetCircumstances cache key: {cacheKey}");
 
             TargetCircumstances targetCircumstances = TargetCircumstancesCache.GetTargetCircumstances(cacheKey);
             if (targetCircumstances == null) {